Sonogram: Change int to size_t

+++ src/analyzer/sonogram.h
@@ -45,7 +45,7 @@ class Sonogram : public Analyzer::Base {
 
  private:
   QPixmap canvas_;
-  int scope_size_;
+  size_t scope_size_;
 };
 
 #endif  // SONOGRAM_H
